Renaming blocks components
* Rename Content component to Blocks component and move it inside blocks folder.
* Rename Block component to BlockItem component with.
