+++ frontend/js/components/blocks/BlockItem.vue
